requirements.txt
flask
requests
python-dotenv
gunicorn
telebot
openai
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
pandas